// src/app/gateway/models/client.Dto.ts
// @ts-nocheck
export  class ClientDto{
    endpointId?: number;
    //endpointId
    clientID?: number;
    clientId?: number;
    //clientId
    name?: string;
    apikey?: string;
    mobileNo?: string;
    allowedAccountno?: number;
    publicKey?: string;
    digitalPublickey?: string;
    status?: number;
    organizationCode?: string;
    shenase?: string;
}